+++ Source/Examples/AvaloniaDemo/MainWindow.axaml.cs
@@ -1,0 +1,17 @@
+// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="MainWindow.axaml.cs" company="PropertyTools">
+//   Copyright (c) 2014 PropertyTools contributors
+// </copyright>
+// --------------------------------------------------------------------------------------------------------------------
+
+using Avalonia.Controls;
+
+namespace AvaloniaDemo;
+
+public partial class MainWindow : Window
+{
+    public MainWindow()
+    {
+        InitializeComponent();
+    }
+}